[FIX] Permission Issues
