+++ taxsystem/models/general.py
@@ -10,8 +10,7 @@ class General(models.Model):
     class Meta:
         managed = False
         permissions = (
-            ("basic_access", _("Can access this app")),
-            ("manage_access", _("Can access tools Tax System")),
+            ("basic_access", _("Can access the Tax System")),
             ("create_access", _("Can add Corporation")),
             ("manage_own_corp", _("Can manage own Corporation")),
             ("manage_corps", _("Can manage all Corporations")),
